% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{ageGroupDoc}
\alias{ageGroupDoc}
\title{Helper for consistent documentation of \code{ageGroup}.}
\arguments{
\item{ageGroup}{If not NULL, a list of ageGroup vectors of length two.}
}
\description{
Helper for consistent documentation of \code{ageGroup}.
}
\keyword{internal}
