% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableConceptClassIds.R
\name{availableConceptClassIds}
\alias{availableConceptClassIds}
\title{Get the available concept classes used in a given set of domains}
\usage{
availableConceptClassIds(cdm, standardConcept = "Standard", domain = NULL)
}
\arguments{
\item{cdm}{A cdm reference to an OMOP CDM dataset. If data is held within a
database, the vocabulary tables should be in the same schema as the clinical
tables (person, observation period, and so on).}

\item{standardConcept}{Character vector with one or more of "Standard",
"Classification", and "Non-standard". These correspond to the flags used
for the standard_concept field in the concept table of the cdm.}

\item{domain}{Character vector with one or more of the OMOP CDM domains.
The results will be restricted to the given domains. Check the available
domains in the database by running 'availableDomains()', or 'associatedDomains()'
to explore the domains associated with a codelist. If NULL, all supported domains
are included: Condition, Drug, Procedure, Device, Observation, and Measurement.}
}
\value{
The concept classes
}
\description{
Get the available concept classes used in a given set of domains
}
\examples{
\donttest{
library(CodelistGenerator)
library(omock)

# Create CDM object
cdm <- mockCdmFromDataset(datasetName = "GiBleed")

# Get all available concept_class_ids in the CDM
availableConceptClassIds(cdm,
                         standardConcept = "Standard")

# Get all available concept_class_ids in the CDM for a specific domain
availableConceptClassIds(cdm,
                         standardConcept = "Standard",
                         domain = "Condition")

# Notice that this corresponds to the information provided by `concept_class_id`
# column in the `concept` table
}
}
