% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{availableIngredients}
\alias{availableIngredients}
\title{Get the names of all available drug ingredients}
\usage{
availableIngredients(cdm)
}
\arguments{
\item{cdm}{A cdm reference via CDMConnector.}
}
\value{
A vector containing the concept names for all ingredient level codes
found in the concept table of cdm.
}
\description{
Get the names of all available drug ingredients
}
\examples{
\donttest{
cdm <- mockVocabRef()
availableIngredients(cdm)
}
}
