% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkCodelistGenerator.R
\name{benchmarkCodelistGenerator}
\alias{benchmarkCodelistGenerator}
\title{Run benchmark of codelistGenerator analyses}
\usage{
benchmarkCodelistGenerator(cdm)
}
\arguments{
\item{cdm}{a CDM reference object}
}
\value{
a tibble with time taken for different analysis
}
\description{
Run benchmark of codelistGenerator analyses
}
\examples{
\donttest{
library(CodelistGenerator)

cdm <- mockVocabRef()

timings <- benchmarkCodelistGenerator(cdm)
}

}
