% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{byConceptDoc}
\alias{byConceptDoc}
\title{Helper for consistent documentation of \code{byConcept}.}
\arguments{
\item{byConcept}{TRUE or FALSE. If TRUE code use will be summarised by
concept.}
}
\description{
Helper for consistent documentation of \code{byConcept}.
}
\keyword{internal}
