% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{bySexDoc}
\alias{bySexDoc}
\title{Helper for consistent documentation of \code{bySex}.}
\arguments{
\item{bySex}{TRUE or FALSE. If TRUE code use will be summarised by sex.}
}
\description{
Helper for consistent documentation of \code{bySex}.
}
\keyword{internal}
