% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{byYearDoc}
\alias{byYearDoc}
\title{Helper for consistent documentation of \code{byYear}.}
\arguments{
\item{byYear}{TRUE or FALSE. If TRUE code use will be summarised by year.}
}
\description{
Helper for consistent documentation of \code{byYear}.
}
\keyword{internal}
