% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{cdmDoc}
\alias{cdmDoc}
\title{Helper for consistent documentation of \code{cdm}.}
\arguments{
\item{cdm}{A cdm reference via CDMConnector.}
}
\description{
Helper for consistent documentation of \code{cdm}.
}
\keyword{internal}
