% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkDbType}
\alias{checkDbType}
\title{Check the database type.}
\usage{
checkDbType(cdm, type = "cdm_reference", messageStore)
}
\arguments{
\item{cdm}{cdm reference}

\item{type}{type of the database, default cdm_reference}

\item{messageStore}{checkmate collection}
}
\description{
Check the database type.
}
