% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkTableExists}
\alias{checkTableExists}
\title{Check if given table exists in cdm.}
\usage{
checkTableExists(cdm, tableName, messageStore)
}
\arguments{
\item{cdm}{the db reference}

\item{tableName}{checkmate collection}

\item{messageStore}{the message store}
}
\description{
Check if given table exists in cdm.
}
