% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{codelistNameDoc}
\alias{codelistNameDoc}
\title{Helper for consistent documentation of \code{codelistNameDoc}}
\arguments{
\item{codelistName}{Name or names of codelist in x. If NULL, all codelist present
in x will be considered.}
}
\description{
Helper for consistent documentation of \code{codelistNameDoc}
}
\keyword{internal}
