% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{countByDoc}
\alias{countByDoc}
\title{Helper for consistent documentation of \code{countBy}.}
\arguments{
\item{countBy}{Either "record" for record-level counts or "person" for
person-level counts.}
}
\description{
Helper for consistent documentation of \code{countBy}.
}
\keyword{internal}
