% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doseFormToRoute}
\alias{doseFormToRoute}
\title{Equivalence from dose from concept IDs to route categories.}
\format{
A data frame with two variables: \code{dose_form_concept_id} and
\code{route_category}.
}
\usage{
doseFormToRoute
}
\description{
Equivalence from dose from concept IDs to route categories.
}
\keyword{datasets}
