% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{doseUnitDoc}
\alias{doseUnitDoc}
\title{Helper for consistent documentation of \code{doseUnit}.}
\arguments{
\item{doseUnit}{Only codes with the specified dose unit
will be returned. If NULL, descendant codes will be returned regardless
of dose unit Use 'availableDoseUnits()' to see the available dose units,
or 'associatedDoseUnits()' to see the associated dose forms in a codelist.}
}
\description{
Helper for consistent documentation of \code{doseUnit}.
}
\keyword{internal}
