% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{.optionsDoc}
\alias{.optionsDoc}
\title{Helper for consistent documentation of \code{.options}.}
\arguments{
\item{.options}{Named list with additional formatting options.
visOmopResults::tableOptions() shows allowed arguments and
their default values.}
}
\description{
Helper for consistent documentation of \code{.options}.
}
\keyword{internal}
