% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCodeSearch.R
\name{emptyCodeSearch}
\alias{emptyCodeSearch}
\title{Empty code search object}
\usage{
emptyCodeSearch(searchStrategy = NULL)
}
\arguments{
\item{searchStrategy}{The search strategy used to generate the candidate
codes.}
}
\value{
An empty \code{code_search} object.
}
\description{
Empty code search object
}
