% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mock_vocab_db.R
\name{generate_mock_vocab_db}
\alias{generate_mock_vocab_db}
\title{Generate example vocabulary database}
\usage{
generate_mock_vocab_db()
}
\value{
DBIConnection to SQLite database
with mock vocabulary
}
\description{
Generate example vocabulary database
}
\examples{
\dontrun{
library(DBI)
library(RSQLite)
library(CodelistGenerator)
db <- generate_mock_vocab_db()
}
}
