% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidate_codes.R
\name{get_candidate_codes}
\alias{get_candidate_codes}
\title{Get candidate codelist}
\usage{
get_candidate_codes(
  keywords,
  domains = c("Condition", "Drug", "Device", "Observation", "Procedure"),
  search.synonyms = FALSE,
  fuzzy.match = FALSE,
  fuzzy.match.max.distance = 0.1,
  exclude = NULL,
  include.descendants = TRUE,
  include.ancestor = FALSE,
  concept,
  concept_synonym,
  concept_ancestor
)
}
\arguments{
\item{keywords}{Character vector}

\item{domains}{Character vector}

\item{search.synonyms}{TRUE/ FALSE}

\item{fuzzy.match}{TRUE/ FALSE}

\item{fuzzy.match.max.distance}{numeric}

\item{exclude}{Character vector}

\item{include.descendants}{TRUE/ FALSE}

\item{include.ancestor}{TRUE/ FALSE}

\item{concept}{Dataframe}

\item{concept_synonym}{Dataframe}

\item{concept_ancestor}{Dataframe}
}
\value{
Dataframe
}
\description{
Get candidate codelist
}
\examples{
library(dplyr)
library(Eunomia)
library(stringr)
library(readr)
connection <- connect(getEunomiaConnectionDetails())
concepts<-querySql(connection, "SELECT * FROM concept;")
concept_ancestors<-querySql(connection, "SELECT * FROM concept_ancestor;")
concept_synonyms<-querySql(connection, "SELECT * FROM concept_synonym;")
disconnect(connection)
names(concepts)<-str_to_lower(names(concepts))
names(concept_ancestors)<-str_to_lower(names(concept_ancestors))
names(concept_synonyms)<-str_to_lower(names(concept_synonyms))
get_candidate_codes(keywords="asthma",
                 concept=concepts,
                    concept_ancestor = concept_ancestors,
                    concept_synonym = concept_synonyms)

}
