% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vocab_version.R
\name{get_vocab_version}
\alias{get_vocab_version}
\title{Get vocabulary version}
\usage{
get_vocab_version(db, vocabulary_database_schema)
}
\arguments{
\item{db}{Database connection via DBI::dbConnect()}

\item{vocabulary_database_schema}{Name of database schema with vocab tables}
}
\value{
Character vector
}
\description{
Get vocabulary version
}
\examples{
library(Eunomia)
library(DBI)
library(RSQLite)
\dontrun{
untar(xzfile(system.file("sqlite", "cdm.tar.xz",
  package = "Eunomia"
), open = "rb"),
exdir = tempdir()
)
db <- DBI::dbConnect(RSQLite::SQLite(), paste0(tempdir(), "\\\\cdm.sqlite"))
get_vocab_version(
  db = db,
  vocabulary_database_schema = "main"
)
}
}
