% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDoseUnit.R
\name{getDoseUnit}
\alias{getDoseUnit}
\title{Get available dose units}
\usage{
getDoseUnit(cdm)
}
\arguments{
\item{cdm}{A cdm reference via CDMConnector.}
}
\value{
A character vector with available routes.
}
\description{
Get the dose form categories available in the database (see
https://doi.org/10.1002/pds.5809 for more details on how routes
were classified).
}
\examples{
\donttest{
library(CodelistGenerator)

cdm <- mockVocabRef()

getDoseUnit(cdm)

}
}
