% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMappings.R
\name{getMappings}
\alias{getMappings}
\title{Show mappings from non-standard vocabularies to standard}
\usage{
getMappings(
  candidateCodelist,
  cdm = NULL,
  nonStandardVocabularies = c("ATC", "ICD10CM", "ICD10PCS", "ICD9CM", "ICD9Proc",
    "LOINC", "OPCS4", "Read", "RxNorm", "RxNorm Extension", "SNOMED")
)
}
\arguments{
\item{candidateCodelist}{Dataframe}

\item{cdm}{cdm_reference via CDMConnector::cdm_from_con()}

\item{nonStandardVocabularies}{Character vector}
}
\value{
tibble
}
\description{
Show mappings from non-standard vocabularies to standard
}
\examples{
\dontrun{
library(DBI)
library(CodelistGenerator)
db <- DBI::dbConnect(" Your database connection here ")
vocabularyDatabaseSchema <- " Your vocabulary schema here "
asthma_codes <- get_candidate_codes(
  keywords = "asthma",
  db = db,
  vocabularyDatabaseSchema = " Your vocabulary schema here "
)
getMappings(
  candidateCodelist = asthma_codes,
  db = db,
  vocabularyDatabaseSchema = " Your vocabulary schema here "
)
}
}
