% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoutes.R
\name{getRoutes}
\alias{getRoutes}
\title{Get available routes in a cdm reference.}
\usage{
getRoutes(cdm, category = TRUE)
}
\arguments{
\item{cdm}{A cdm reference.}

\item{category}{Whether to use the derived dose form categories (see
https://doi.org/10.1002/pds.5809) or the routes from vocabulary tables.}
}
\value{
A character vector with available routes
}
\description{
Get available routes in a cdm reference.
}
