% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{getVocabularies}
\alias{getVocabularies}
\title{Get the vocabularies available in the cdm}
\usage{
getVocabularies(cdm)
}
\arguments{
\item{cdm}{A cdm reference via CDMConnector.}
}
\value{
Names of available vocabularies.
}
\description{
Get the vocabularies available in the cdm
}
\examples{
\donttest{
cdm <- mockVocabRef()
getVocabularies(cdm = cdm)
}
}
