% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{getVocabVersion}
\alias{getVocabVersion}
\title{Get the version of the vocabulary used in the cdm}
\usage{
getVocabVersion(cdm)
}
\arguments{
\item{cdm}{A cdm reference via CDMConnector.}
}
\value{
The vocabulary version being used in the cdm.
}
\description{
Get the version of the vocabulary used in the cdm
}
\examples{
\donttest{
cdm <- mockVocabRef()
getVocabVersion(cdm = cdm)
}
}
