% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{groupColumnDoc}
\alias{groupColumnDoc}
\title{Helper for consistent documentation of \code{groupColumn}.}
\arguments{
\item{groupColumn}{Variables to use as group labels. Allowed columns are:
"cdm_name", "codelist_name", "domain_id", "standard_concept_name",
"standard_concept_id", "estimate_name", "standard_concept", "vocabulary_id".
These cannot be used in header.}
}
\description{
Helper for consistent documentation of \code{groupColumn}.
}
\keyword{internal}
