% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{groupColumnStrataDoc}
\alias{groupColumnStrataDoc}
\title{Helper for consistent documentation of \code{groupColumn}.}
\arguments{
\item{groupColumn}{Variables to use as group labels. Allowed columns are:
"cdm_name", "codelist_name", "standard_concept_name", "standard_concept_id",
"estimate_name", "source_concept_name", "source_concept_id", "domain_id". If
results are stratified, "year", "sex", "age_group" can also be used.
These cannot be used in header.}
}
\description{
Helper for consistent documentation of \code{groupColumn}.
}
\keyword{internal}
