% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{hideDoc}
\alias{hideDoc}
\title{Helper for consistent documentation of \code{hide}.}
\arguments{
\item{hide}{Table columns to exclude, options are:  "cdm_name",
"codelist_name", "domain_id", "standard_concept_name", "standard_concept_id",
"estimate_name", "standard_concept", "vocabulary_id". These cannot be used in
header or groupColumn.}
}
\description{
Helper for consistent documentation of \code{hide}.
}
\keyword{internal}
