% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{hideStrataDoc}
\alias{hideStrataDoc}
\title{Helper for consistent documentation of \code{hide}.}
\arguments{
\item{hide}{Table columns to exclude, options are: "cdm_name",
"codelist_name", "year", "sex", "age_group", "standard_concept_name",
"standard_concept_id", "estimate_name", "source_concept_name",
"source_concept_id", "domain_id". If results are stratified, "year", "sex",
"age_group" can also be used. These cannot be used in header or groupColumn.}
}
\description{
Helper for consistent documentation of \code{hide}.
}
\keyword{internal}
