% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{includeDescendantsDoc}
\alias{includeDescendantsDoc}
\title{Helper for consistent documentation of \code{includeDescendants}.}
\arguments{
\item{includeDescendants}{Either TRUE or FALSE. If TRUE descendant concepts
of identified concepts will be included in the candidate codelist. If FALSE
only direct mappings from ICD-10 codes to standard codes will be returned.}
}
\description{
Helper for consistent documentation of \code{includeDescendants}.
}
\keyword{internal}
