% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{ingredientRangeDoc}
\alias{ingredientRangeDoc}
\title{Helper for consistent documentation of \code{ingredientRange}.}
\arguments{
\item{ingredientRange}{Used to restrict descendant codes to those
associated with a specific number of drug ingredients. Must be a vector of
length two with the first element the minimum number of ingredients allowed
and the second the maximum. A value of c(2, 2) would restrict to only
concepts associated with two ingredients.}
}
\description{
Helper for consistent documentation of \code{ingredientRange}.
}
\keyword{internal}
