% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{levelATCDoc}
\alias{levelATCDoc}
\title{Helper for consistent documentation of \code{level}.}
\arguments{
\item{level}{ATC level. Can be one or more of "ATC 1st", "ATC 2nd",
"ATC 3rd", "ATC 4th", and "ATC 5th".}
}
\description{
Helper for consistent documentation of \code{level}.
}
\keyword{internal}
