% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{levelICD10Doc}
\alias{levelICD10Doc}
\title{Helper for consistent documentation of \code{level}.}
\arguments{
\item{level}{Can be either "ICD10 Chapter", "ICD10 SubChapter",
"ICD10 Hierarchy", or "ICD10 Code".}
}
\description{
Helper for consistent documentation of \code{level}.
}
\keyword{internal}
