% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{minimumCountDoc}
\alias{minimumCountDoc}
\title{Helper for consistent documentation of \code{minimumCount}.}
\arguments{
\item{minimumCount}{Any codes with a frequency under this will be removed.}
}
\description{
Helper for consistent documentation of \code{minimumCount}.
}
\keyword{internal}
