% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockVocab.R
\name{mockVocab}
\alias{mockVocab}
\title{Generate example vocabulary database}
\usage{
mockVocab(dbType = "SQLite")
}
\arguments{
\item{dbType}{Character vector with type of in memory database.
Either 'SQL lite' or 'duckdb'}
}
\value{
DBIConnection to SQLite database
with mock vocabulary
}
\description{
Generate example vocabulary database
}
\examples{
\dontrun{
library(DBI)
library(RSQLite)
library(CodelistGenerator)
db <- mockVocab()
}
}
