% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{nameStyleDoc}
\alias{nameStyleDoc}
\title{Helper for consistent documentation of \code{nameStyle}.}
\arguments{
\item{nameStyle}{Name style to apply to returned list. Can be one of
\code{"{concept_code}"},\code{"{concept_id}"}, \code{"{concept_name}"}, or a combination (i.e.,
\code{"{concept_code}_{concept_name}"}).}
}
\description{
Helper for consistent documentation of \code{nameStyle}.
}
\keyword{internal}
