% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConcepts.R
\name{readConceptList}
\alias{readConceptList}
\title{Get concept ids from a provided path to json files}
\usage{
readConceptList(path, cdm)
}
\arguments{
\item{path}{path to a file or folder containing jsons to be read}

\item{cdm}{A cdm reference created with CDMConnector}
}
\value{
list of concept_ids and respective concept_ids of interest
}
\description{
Get concept ids from a provided path to json files
}
