% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cohortCodelist}
\alias{importCodelist}
\alias{importConceptSetExpression}
\alias{exportCodelist}
\alias{exportConceptSetExpression}
\alias{newCodelist}
\alias{newCodelistWithDetails}
\alias{newConceptSetExpression}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{omopgenerics}{\code{\link[omopgenerics]{cohortCodelist}}, \code{\link[omopgenerics]{exportCodelist}}, \code{\link[omopgenerics]{exportConceptSetExpression}}, \code{\link[omopgenerics]{importCodelist}}, \code{\link[omopgenerics]{importConceptSetExpression}}, \code{\link[omopgenerics]{newCodelist}}, \code{\link[omopgenerics]{newCodelistWithDetails}}, \code{\link[omopgenerics]{newConceptSetExpression}}}
}}

