% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{routeCategoryDoc}
\alias{routeCategoryDoc}
\title{Helper for consistent documentation of \code{routeCategory}.}
\arguments{
\item{routeCategory}{Only codes with the specified route will be
returned. If NULL, descendant codes will be returned regardless of route
category. Use 'availableRouteCategories()' to find the available route categories in
the database, and 'associatedRouteCategories()' to get drug routs associated with
a codelist.}
}
\description{
Helper for consistent documentation of \code{routeCategory}.
}
\keyword{internal}
