% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_mappings.R
\name{show_mappings}
\alias{show_mappings}
\title{Show mappings from source vocabularies to standard}
\usage{
show_mappings(
  candidate_codelist,
  source_vocabularies = c("ATC", "ICD10CM", "ICD10PCS", "ICD9CM", "ICD9Proc", "LOINC",
    "OPCS4", "Read", "RxNorm", "RxNorm Extension", "SNOMED"),
  concept,
  concept_relationship
)
}
\arguments{
\item{candidate_codelist}{Dataframe}

\item{source_vocabularies}{Character vector}

\item{concept}{Dataframe}

\item{concept_relationship}{Dataframe}
}
\value{
Dataframe
}
\description{
Show mappings from source vocabularies to standard
}
\examples{
# note, Eunomia, which is used for the example below, does not include a full set of vocabularies. The full set can be downloaded from https://athena.ohdsi.org
library(dplyr)
library(Eunomia)
library(stringr)
library(readr)
connection <- connect(getEunomiaConnectionDetails())
concepts<-querySql(connection, "SELECT * FROM concept;")
concept_relationships<-querySql(connection, "SELECT * FROM CONCEPT_RELATIONSHIP;")
concept_ancestors<-querySql(connection, "SELECT * FROM concept_ancestor;")
concept_synonyms<-querySql(connection, "SELECT * FROM concept_synonym;")
disconnect(connection)
names(concepts)<-str_to_lower(names(concepts))
names(concept_relationships)<-str_to_lower(names(concept_relationships))
names(concept_ancestors)<-str_to_lower(names(concept_ancestors))
names(concept_synonyms)<-str_to_lower(names(concept_synonyms))
asthma_codes<- get_candidate_codes(keywords="asthma",
                 concept=concepts,
                    concept_ancestor = concept_ancestors,
                    concept_synonym = concept_synonyms)
show_mappings(candidate_codelist= asthma_codes,
           concept = concepts,
           concept_relationship =  concept_relationships)
}
