% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMappings.R
\name{showMappings}
\alias{showMappings}
\title{Show mappings from non-standard vocabularies to standard}
\usage{
showMappings(
  candidateCodelist,
  db = NULL,
  vocabularyDatabaseSchema = NULL,
  arrowDirectory = NULL,
  nonStandardVocabularies = c("ATC", "ICD10CM", "ICD10PCS", "ICD9CM", "ICD9Proc",
    "LOINC", "OPCS4", "Read", "RxNorm", "RxNorm Extension", "SNOMED")
)
}
\arguments{
\item{candidateCodelist}{Dataframe}

\item{db}{Database connection via DBI::dbConnect()}

\item{vocabularyDatabaseSchema}{Name of database
schema with vocab tables}

\item{arrowDirectory}{Path to folder containing output of
Codelist_generator::importVocab() - five parquet files: 'concept',
'concept_ancestor', 'concept_relationship', 'concept_synonym', and
'vocabulary. Required if db is NULL}

\item{nonStandardVocabularies}{Character vector}
}
\value{
tibble
}
\description{
Show mappings from non-standard vocabularies to standard
}
\examples{
\dontrun{
library(DBI)
library(CodelistGenerator)
db <- DBI::dbConnect(" Your database connection here ")
vocabularyDatabaseSchema <- " Your vocabulary schema here "
asthma_codes <- get_candidate_codes(
  keywords = "asthma",
  db = db,
  vocabularyDatabaseSchema = " Your vocabulary schema here "
)
showMappings(
  candidateCodelist = asthma_codes,
  db = db,
  vocabularyDatabaseSchema = " Your vocabulary schema here "
)
}
}
