% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesInUse.R
\name{sourceCodesInUse}
\alias{sourceCodesInUse}
\title{Get the source codes being used in patient records}
\usage{
sourceCodesInUse(
  cdm,
  table = c("condition_occurrence", "device_exposure", "drug_exposure", "measurement",
    "observation", "procedure_occurrence", "visit_occurrence")
)
}
\arguments{
\item{cdm}{A cdm reference via CDMConnector.}

\item{table}{cdm table of interest.}
}
\value{
A list of source codes used in the database.
}
\description{
Get the source codes being used in patient records
}
\examples{
\donttest{
cdm <- mockVocabRef("database")
x <- sourceCodesInUse(cdm = cdm)
x
CDMConnector::cdmDisconnect(cdm)
}
}
