% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{standardConceptDoc}
\alias{standardConceptDoc}
\title{Helper for consistent documentation of \code{standardConcept}.}
\arguments{
\item{standardConcept}{Character vector with one or more of "Standard",
"Classification", and "Non-standard". These correspond to the flags used
for the standard_concept field in the concept table of the cdm.}
}
\description{
Helper for consistent documentation of \code{standardConcept}.
}
\keyword{internal}
