% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{tableDoc}
\alias{tableDoc}
\title{Helper for consistent documentation of \code{table}.}
\arguments{
\item{table}{cdm table of interest.}
}
\description{
Helper for consistent documentation of \code{table}.
}
\keyword{internal}
