% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{typeBroadDoc}
\alias{typeBroadDoc}
\title{Helper for consistent documentation of \code{type}.}
\arguments{
\item{type}{Can be "codelist", "codelist_with_details" or
"concept_set_expression".}
}
\description{
Helper for consistent documentation of \code{type}.
}
\keyword{internal}
