% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{typeNarrowDoc}
\alias{typeNarrowDoc}
\title{Helper for consistent documentation of \code{type}.}
\arguments{
\item{type}{Can be "codelist" or "codelist_with_details".}
}
\description{
Helper for consistent documentation of \code{type}.
}
\keyword{internal}
