% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{typeTableDoc}
\alias{typeTableDoc}
\title{Helper for consistent documentation of \code{type}.}
\arguments{
\item{type}{Type of desired formatted table. By default (type = NULL), it will
create a 'gt' table. To see supported formats use visOmopResults::tableType().}
}
\description{
Helper for consistent documentation of \code{type}.
}
\keyword{internal}
