% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabRef.R
\name{vocabRefFromDatabase}
\alias{vocabRefFromDatabase}
\title{Create a list of references to remote OMOP vocab tables}
\usage{
vocabRefFromDatabase(con, schema = NULL)
}
\arguments{
\item{con}{A database connection}

\item{schema}{The schema where the vocab tables are located. Defaults to NULL.}
}
\value{
A list of dplyr database table references pointing to 5 vocabulary tables
}
\description{
Create a list of references to remote OMOP vocab tables
}
