% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabRef.R
\name{vocabRefFromFiles}
\alias{vocabRefFromFiles}
\title{Create a list of vocab tables from a directory of parquet files}
\usage{
vocabRefFromFiles(path)
}
\arguments{
\item{path}{Directory containing  5 vocabulary parquet files:
"concept", "concept_ancestor", "concept_synonym", "concept_relationship", "vocabulary" all with
the .parquet extension}
}
\value{
A list of 5 Arrow Tables
}
\description{
Create a list of vocab tables from a directory of parquet files
}
