% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{vocabularyVersion}
\alias{vocabularyVersion}
\title{Get the available version of the vocabulary used in the cdm}
\usage{
vocabularyVersion(cdm)
}
\arguments{
\item{cdm}{A cdm reference to an OMOP CDM dataset. If data is held within a
database, the vocabulary tables should be in the same schema as the clinical
tables (person, observation period, and so on).}
}
\value{
The vocabulary version being used in the cdm.
}
\description{
Get the available version of the vocabulary used in the cdm
}
\examples{
\donttest{
library(CodelistGenerator)
cdm <- mockVocabRef()
vocabularyVersion(cdm = cdm)
}
}
