% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{xDoc}
\alias{xDoc}
\title{Helper for consistent documentation of \code{x}.}
\arguments{
\item{x}{A codelist, codelist_with_details, or a concept_set. See \code{newCodelist()}, \code{newCodelistWithDetails()}, \code{newConceptSetExpression()} functions for more details.}
}
\description{
Helper for consistent documentation of \code{x}.
}
\keyword{internal}
