% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{availableATC}
\alias{availableATC}
\title{Get all ATC codes from the cdm}
\usage{
availableATC(cdm, level = c("ATC 1st"))
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{level}{ATC level. Can be one or more of "ATC 1st", "ATC 2nd",
"ATC 3rd", "ATC 4th", and "ATC 5th"}
}
\value{
A vector list of all ATC codes for the chosen level(s) found in the
concept table of cdm.
}
\description{
Get all ATC codes from the cdm
}
\examples{
\dontrun{
cdm <- mockVocabRef()
availableATC(cdm)
}

}
