% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{availableICD10}
\alias{availableICD10}
\title{Get all ICD codes from the cdm}
\usage{
availableICD10(cdm, level = c("ICD10 Chapter", "ICD10 SubChapter"))
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{level}{Can be either "ICD10 Chapter" or "ICD10 SubChapter"}
}
\value{
A vector list of all ICD10 codes for the chosen level(s) found in the
concept table of cdm.
}
\description{
Get all ICD codes from the cdm
}
\examples{
\dontrun{
cdm <- mockVocabRef()
availableICD10(cdm)
}
}
