% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{availableIngredients}
\alias{availableIngredients}
\title{Get all ingredients codes from the cdm}
\usage{
availableIngredients(cdm)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}
}
\value{
A vector list of all ingredient level codes found in the concept
table of cdm.
}
\description{
Get all ingredients codes from the cdm
}
\examples{
\dontrun{
cdm <- mockVocabRef()
availableIngredients(cdm)
}
}
