% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_words.R
\name{clean_words}
\alias{clean_words}
\title{Prepare words for search.}
\usage{
clean_words(words)
}
\arguments{
\item{words}{A character vector.}
}
\value{
A character vector of the same length.
}
\description{
Prepare words for search.
}
\examples{
clean_words(c("Type-1 diabetes "))
}
