% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodelistGenerator-package.R
\docType{package}
\name{CodelistGenerator-package}
\alias{CodelistGenerator}
\alias{CodelistGenerator-package}
\title{CodelistGenerator: Generate Candidate Codelists for the OMOP CDM}
\description{
Generate a candidate codelist for the OMOP CDM based on string matching. For a given search strategy, a candidate codelist will be returned.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu.github.io/CodelistGenerator/}
}

}
\author{
\strong{Maintainer}: Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})

}
\keyword{internal}
