% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodelistGenerator-package.R
\docType{package}
\name{CodelistGenerator-package}
\alias{CodelistGenerator}
\alias{CodelistGenerator-package}
\title{CodelistGenerator: Identify Relevant Clinical Codes and Evaluate Their Use}
\description{
Generate a candidate code list for the Observational Medical Outcomes Partnership (OMOP) common data model based on string matching. For a given search strategy, a candidate code list will be returned.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu.github.io/CodelistGenerator/}
}

}
\author{
\strong{Maintainer}: Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})

Other contributors:
\itemize{
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID}) [contributor]
  \item Xihang  Chen \email{xihang.chen@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0001-8112-8959}{ORCID}) [contributor]
}

}
\keyword{internal}
