% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesFromConceptSet.R
\name{codesFromCohort}
\alias{codesFromCohort}
\title{Get concept ids from a provided path to cohort json files}
\usage{
codesFromCohort(path, cdm)
}
\arguments{
\item{path}{Path to a file or folder containing JSONs of concept sets}

\item{cdm}{A cdm reference created with CDMConnector}
}
\value{
Named list with concept_ids for each concept set
}
\description{
Get concept ids from a provided path to cohort json files
}
