% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesFromConceptSet.R
\name{codesFromConceptSet}
\alias{codesFromConceptSet}
\title{Get concept ids from a provided path to json files}
\usage{
codesFromConceptSet(path, cdm, withConceptDetails = FALSE)
}
\arguments{
\item{path}{Path to a file or folder containing JSONs of concept sets}

\item{cdm}{A cdm reference created with CDMConnector}

\item{withConceptDetails}{If FALSE a vector of concept IDs will be returned
for each concept set. If TRUE a tibble will be returned with additional
information on the identified concepts.}
}
\value{
Named list with concept_ids for each concept set
}
\description{
Get concept ids from a provided path to json files
}
