% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesInUse.R
\name{codesInUse}
\alias{codesInUse}
\title{Get codes used in the database}
\usage{
codesInUse(
  cdm,
  table = c("condition_occurrence", "device_exposure", "drug_exposure", "measurement",
    "observation", "procedure_occurrence", "visit_occurrence")
)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{table}{cdm table}
}
\description{
Get codes used in the database
}
