% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCodelists.R
\name{compareCodelists}
\alias{compareCodelists}
\title{Compare two codelists}
\usage{
compareCodelists(codelist1, codelist2)
}
\arguments{
\item{codelist1}{Output of getCandidateCodes}

\item{codelist2}{Output of getCandidateCodes}
}
\value{
tibble
}
\description{
Compare two codelists
}
\examples{
\dontrun{
library(DBI)
library(CodelistGenerator)
db <- DBI::dbConnect(" Your database connection here ")
vocabularyDatabaseSchema <- " Your vocabulary schema here "
asthmaCodes <- getCandidateCodes(
  keywords = "asthma",
  db = db,
  vocabularyDatabaseSchema = " Your vocabulary schema here "
)
persistantAsthmaCodes <- getCandidateCodes(
  keywords = "Persistent asthma",
  db = db,
  vocabularyDatabaseSchema = " Your vocabulary schema here "
)
compareCodelists(
  codelist1 = asthmaCodes,
  codelist2 = persistantAsthmaCodes
)
}
}
