% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{domainDoc}
\alias{domainDoc}
\title{Helper for consistent documentation of \code{domain}.}
\arguments{
\item{domain}{Character vector with one or more of the OMOP CDM domains.
The results will be restricted to the given domains. Check the available
ones by running getDomains(). If NULL, all supported domains are included:
Condition, Drug, Procedure, Device, Observation, and Measurement.}
}
\description{
Helper for consistent documentation of \code{domain}.
}
\keyword{internal}
