% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doseFormToRoute}
\alias{doseFormToRoute}
\title{Table showing the route category associated with each dose form.}
\format{
A data frame
\describe{
\item{dose_form_concept_id}{Concept ID of each dose form}
\item{dose_form_concept_name}{Concept name of each dose form}
\item{route_category}{Route category associated to the dose form}
}
}
\usage{
doseFormToRoute
}
\description{
Table showing the route category associated with each dose form.
}
\examples{
\donttest{
library(CodelistGenerator)
doseFormToRoute
}
}
\keyword{data}
