% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadVocab.R
\name{downloadVocab}
\alias{downloadVocab}
\title{Download OMOP vocabulary tables}
\usage{
downloadVocab(conn, schema, dirOut, errorIfExists = TRUE, verbose = FALSE)
}
\arguments{
\item{conn}{Database connection via DBI::dbConnect()}

\item{schema}{Name of database schema containing vocab tables}

\item{dirOut}{Directory where output files will be saved}

\item{verbose}{Either TRUE or FALSE. If TRUE, progress messages will be reported.}

\item{errorIfExist}{Either TRUE or FALSE.
If TRUE, an error will be thrown if vocab parquet files already exist.
If FALSE, existing vocab parquet files will be overwritten.}
}
\value{
Invisibly returns TRUE. This function is called for its side effect
of creating parquet files in the specified directory.
}
\description{
Download the OMOP vocabulary tables from a remote database
connection and save them as parquet files in the specified folder.
}
\examples{
\dontrun{
library(CodelistGenerator)
conn <- generateMockVocabDb()
dOut <- here()

downloadVocab(
  conn = conn,
  schema = "main",
  dirOut = dOut
)
}

}
