% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadVocab.R
\name{downloadVocab}
\alias{downloadVocab}
\title{Download vocabulary tables}
\usage{
downloadVocab(
  db,
  vocabularyDatabaseSchema,
  dirOut,
  errorIfExists = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{db}{Database connection via DBI::dbConnect()}

\item{vocabularyDatabaseSchema}{Name of database
schema with vocab tables}

\item{dirOut}{Directory where output files will
be saved}

\item{errorIfExists}{Either TRUE or FALSE.
If TRUE, error thrown if output database already exists.
If FALSE, output database overwritten if it exists.}

\item{verbose}{Either TRUE or FALSE.
If TRUE, progress will be reported.}
}
\value{
Apache arrow parquet files
}
\description{
Download vocabulary tables
}
\examples{
\dontrun{
library(DBI)
library(Arrow)
library(CodelistGenerator)
library(here)
db <- generateMockVocabDb()
dOut <- here()

downloadVocab(
  db = db,
  vocabularyDatabaseSchema = "main",
  dirOut = dOut
)
}

}
