% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMockVocabDb.R
\name{generateMockVocabDb}
\alias{generateMockVocabDb}
\title{Generate example vocabulary database}
\usage{
generateMockVocabDb()
}
\value{
DBIConnection to SQLite database
with mock vocabulary
}
\description{
Generate example vocabulary database
}
\examples{
\dontrun{
library(DBI)
library(RSQLite)
library(CodelistGenerator)
db <- generateMockVocabDb()
}
}
