% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidate_codes.R
\name{get_candidate_codes}
\alias{get_candidate_codes}
\title{Generate candidate codelist for the OMOP CDM}
\usage{
get_candidate_codes(
  keywords,
  domains = "Condition",
  standard_concept = "Standard",
  search_synonyms = FALSE,
  search_source = FALSE,
  fuzzy_match = FALSE,
  max_distance_cost = 0.1,
  exclude = NULL,
  include_descendants = TRUE,
  include_ancestor = FALSE,
  verbose = FALSE,
  db,
  vocabulary_database_schema
)
}
\arguments{
\item{keywords}{Character vector of words to search for.
Where more than one word is given (e.g. "knee osteoarthritis"),
all combinations of those words should be identified
positions (e.g. "osteoarthritis of knee") should be identified.}

\item{domains}{Character vector with one or more of the OMOP CDM domain.}

\item{standard_concept}{Character vector with one or more of "Standard",
"Classification", and "Non-standard". These correspond to the flags used
for the standard_concept field in the concept table of the cdm.}

\item{search_synonyms}{Either TRUE or FALSE. If TRUE the code will also
search via the concept synonym table.}

\item{search_source}{Either TRUE or FALSE. If TRUE the code will also
search via source concepts.}

\item{fuzzy_match}{Either TRUE or FALSE. If TRUE the fuzzy matching
will be used, with approximate matches identified.}

\item{max_distance_cost, }{The
maximum number/fraction of match cost (generalized Levenshtein distance)
for fuzzy matching (see ??base::agrep for further details).}

\item{exclude}{Character vector of words
to identify concepts to exclude.}

\item{include_descendants}{Either TRUE or FALSE.
If TRUE descendant concepts of identified concepts
will be included in the candidate codelist.}

\item{include_ancestor}{Either TRUE or FALSE.
If TRUE the direct ancestor concepts of identified concepts
will be included in the candidate codelist.}

\item{verbose}{Either TRUE or FALSE.
If TRUE, progress will be reported.}

\item{db}{Database connection via DBI::dbConnect()}

\item{vocabulary_database_schema}{Name of database
schema with vocab tables}
}
\value{
tibble
}
\description{
This function generates a set of codes that
can be considered for creating a phenotype
using the OMOP CDM.
}
\examples{
\dontrun{
library(DBI)
library(CodelistGenerator)
db <- DBI::dbConnect(" Your database connection here " )
vocabulary_database_schema <- " Your vocabulary schema here "
get_candidate_codes(
  keywords = "asthma",
  db = db,
  vocabulary_database_schema = vocabulary_database_schema
)
}
}
