% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidate_codes.R
\name{get_candidate_codes}
\alias{get_candidate_codes}
\title{Generate candidate codelist for the OMOP CDM}
\usage{
get_candidate_codes(
  keywords,
  domains = c("Condition", "Drug", "Device", "Observation", "Procedure"),
  search.synonyms = FALSE,
  fuzzy.match = FALSE,
  fuzzy.match.max.distance = 0.1,
  exclude = NULL,
  include.descendants = TRUE,
  include.ancestor = FALSE,
  db,
  vocabulary_database_schema
)
}
\arguments{
\item{keywords}{Character vector of words to search for. Where more than one word is given (e.g. "knee osteoarthritis"), all words will be identified but can be in different positions (e.g. "osteoarthritis of knee") should be identified.}

\item{domains}{Character vector with one or more of the OMOP CDM domain (e.g. "Condition").}

\item{search.synonyms}{Either TRUE or FALSE. If TRUE the code will also search via the concept synonym table.}

\item{fuzzy.match}{Either TRUE or FALSE. If TRUE the fuzzy matches will be used, with approximate matches identified.}

\item{fuzzy.match.max.distance}{The max.distance parmeter for fuzzy matching (see ??base::agrep for further details).}

\item{exclude}{Character vector of words to search for to identify concepts to exclude.}

\item{include.descendants}{Either TRUE or FALSE. If TRUE descendant concepts of identified concepts will be included in the candidate codelist.}

\item{include.ancestor}{Either TRUE or FALSE. If TRUE the direct ancestor concepts of identified concepts will be included in the candidate codelist.}

\item{db}{Database connection via DBI::dbConnect()}

\item{vocabulary_database_schema}{Name of database schema with vocab tables}
}
\value{
Dataframe
}
\description{
This function generates a set of codes that can be considered for creating a phenotype
using the OMOP CDM.
}
\examples{
### note, Eunomia, used in the example below, does not include a full set of vocabularies. The full set can be downloaded from https://athena.ohdsi.org
library(Eunomia)
library(DBI)
library(RSQLite)
untar(xzfile(system.file("sqlite", "cdm.tar.xz", package = "Eunomia"), open = "rb"),
       exdir =  tempdir())
db <- DBI::dbConnect(RSQLite::SQLite(), paste0(tempdir(),"\\\\cdm.sqlite"))
get_candidate_codes(keywords="asthma",
                   search.synonyms=TRUE,
                             fuzzy.match=TRUE,
                             exclude=NULL,
                             include.descendants=TRUE,
                             include.ancestor=FALSE,
                   db=db,
                   vocabulary_database_schema = "main")

}
