% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCandidateCodes.R
\name{getCandidateCodes}
\alias{getCandidateCodes}
\title{Generate candidate codelist for the OMOP CDM}
\usage{
getCandidateCodes(
  keywords,
  db = NULL,
  vocabularyDatabaseSchema = NULL,
  arrowDirectory = NULL,
  exclude = NULL,
  domains = "Condition",
  conceptClassId = NULL,
  standardConcept = "Standard",
  searchViaSynonyms = FALSE,
  searchNonStandard = FALSE,
  fuzzyMatch = FALSE,
  maxDistanceCost = 0.1,
  includeDescendants = TRUE,
  includeAncestor = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{keywords}{Character vector of words to search for.
Where more than one word is given (e.g. "knee osteoarthritis"),
all combinations of those words should be identified
positions (e.g. "osteoarthritis of knee") should be identified.}

\item{db}{Database connection via DBI::dbConnect(). Required if
arrowDirectory is NULL}

\item{vocabularyDatabaseSchema}{Name of database
schema with vocab tables}

\item{arrowDirectory}{Path to folder containing output of
Codelist_generator::importVocab() - five parquet files: 'concept',
'concept_ancestor', 'concept_relationship', 'concept_synonym', and
'vocabulary. Required if db is NULL}

\item{exclude}{Character vector of words
to identify concepts to exclude.}

\item{domains}{Character vector with one or more of the OMOP CDM domain.}

\item{conceptClassId}{Character vector with one or more concept class
of the Concept}

\item{standardConcept}{Character vector with one or more of "Standard",
"Classification", and "Non-standard". These correspond to the flags used
for the standard_concept field in the concept table of the cdm.}

\item{searchViaSynonyms}{Either TRUE or FALSE. If TRUE the code will also
search via the concept synonym table.}

\item{searchNonStandard}{Either TRUE or FALSE. If TRUE the code will also
search via non-standard concepts.}

\item{fuzzyMatch}{Either TRUE or FALSE. If TRUE the fuzzy matching
will be used, with approximate matches identified.}

\item{maxDistanceCost, }{The
maximum number/fraction of match cost (generalized Levenshtein distance)
for fuzzy matching (see ??base::agrep for further details).}

\item{includeDescendants}{Either TRUE or FALSE.
If TRUE descendant concepts of identified concepts
will be included in the candidate codelist.}

\item{includeAncestor}{Either TRUE or FALSE.
If TRUE the direct ancestor concepts of identified concepts
will be included in the candidate codelist.}

\item{verbose}{Either TRUE or FALSE.
If TRUE, progress will be reported.}
}
\value{
tibble
}
\description{
This function generates a set of codes that
can be considered for creating a phenotype
using the OMOP CDM.
}
\examples{
\dontrun{
library(DBI)
library(CodelistGenerator)
db <- DBI::dbConnect(" Your database connection here ")
vocabularyDatabaseSchema <- " Your vocabulary schema here "
getCandidateCodes(
  keywords = "asthma",
  db = db,
  vocabularyDatabaseSchema = vocabularyDatabaseSchema
)
}
}
