% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{getconceptClassId}
\alias{getconceptClassId}
\title{getconceptClassId}
\usage{
getconceptClassId(cdm, domain = NULL)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{domain}{Vocabulary domain}
}
\description{
getconceptClassId
}
