% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{getConceptClassId}
\alias{getConceptClassId}
\title{getConceptClassId}
\usage{
getConceptClassId(cdm, standardConcept = "Standard", domain = NULL)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{standardConcept}{Character vector with one or more of "Standard",
"Classification", and "Non-standard". These correspond to the flags used
for the standard_concept field in the concept table of the cdm.}

\item{domain}{Vocabulary domain}
}
\value{
The concept class used for a given set of domains
}
\description{
getConceptClassId
}
\examples{
cdm <- mockVocabRef()
getConceptClassId(cdm = cdm, domain = "drug")
DBI::dbDisconnect(attr(cdm, "dbcon"), shutdown = TRUE)
}
