% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{getDescendants}
\alias{getDescendants}
\title{getDescendants}
\usage{
getDescendants(cdm, concept_id)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{concpet_id}{concpet_id to search}
}
\value{

}
\description{
getDescendants
}
