% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{getDescendants}
\alias{getDescendants}
\title{getDescendants}
\usage{
getDescendants(cdm, conceptId)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{conceptId}{concpet_id to search}
}
\value{
The descendants of a given concept id
}
\description{
getDescendants
}
\examples{
cdm <- mockVocabRef()
getDescendants(cdm = cdm, conceptId = 1)
DBI::dbDisconnect(attr(cdm, "dbcon"), shutdown = TRUE)
}
