% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{getDomains}
\alias{getDomains}
\title{getDomains}
\usage{
getDomains(cdm, standardConcept = "Standard")
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{standardConcept}{Character vector with one or more of "Standard",
"Classification", and "Non-standard". These correspond to the flags used
for the standard_concept field in the concept table of the cdm.}
}
\value{
The domains of the cdm
}
\description{
getDomains
}
\examples{
\donttest{
cdm <- mockVocabRef()
getDomains(cdm = cdm)
}
}
