% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICD10StandardCodes.R
\name{getICD10StandardCodes}
\alias{getICD10StandardCodes}
\title{Get corresponding standard codes for ICD-10 chapters and sub-chapters}
\usage{
getICD10StandardCodes(
  cdm,
  level = c("ICD10 Chapter", "ICD10 SubChapter"),
  name = NULL,
  includeDescendants = TRUE,
  withConceptDetails = FALSE
)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{level}{Can be either "ICD10 Chapter" or "ICD10 SubChapter"}

\item{name}{Name of chapter or sub-chapter of interest. If NULL, all
will be considered.}

\item{includeDescendants}{If FALSE only direct mappings from ICD-10 codes
to standard codes will be returned. If TRUE descendants of standard concepts
will also be included.}

\item{withConceptDetails}{If FALSE a vector of concept IDs will be returned
for each ICD group If TRUE a tibble will be returned with additional
information on the identified concepts.}
}
\value{
A named list, with each element containing the corresponding
standard codes (and descendants) of ICD chapters and sub-chapters
}
\description{
Get corresponding standard codes for ICD-10 chapters and sub-chapters
}
\examples{
\dontrun{
cdm <- mockVocabRef()
getICD10StandardCodes(cdm = cdm, level = c(
  "ICD10 Chapter",
  "ICD10 SubChapter"
))
CDMConnector::cdmDisconnect(cdm)
}
}
