% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICD10StandardCodes.R
\name{getICD10StandardCodes}
\alias{getICD10StandardCodes}
\title{Get corresponding standard codes for ICD-10 chapters and sub-chapters}
\usage{
getICD10StandardCodes(
  cdm,
  level = c("ICD10 Chapter", "ICD10 SubChapter"),
  name = NULL,
  includeDescendants = TRUE,
  type = "codelist"
)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{level}{Can be either "ICD10 Chapter" or "ICD10 SubChapter"}

\item{name}{Name of chapter or sub-chapter of interest. If NULL, all
will be considered.}

\item{includeDescendants}{If FALSE only direct mappings from ICD-10 codes
to standard codes will be returned. If TRUE descendants of standard concepts
will also be included.}

\item{type}{Can be "codelist", "codelist_with_details", or
"concept_set_expression"}
}
\value{
A named list, with each element containing the corresponding
standard codes (and descendants) of ICD chapters and sub-chapters
}
\description{
Get corresponding standard codes for ICD-10 chapters and sub-chapters
}
\examples{
\donttest{
cdm <- mockVocabRef()
getICD10StandardCodes(cdm = cdm, level = c(
  "ICD10 Chapter",
  "ICD10 SubChapter"
))
}
}
