% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{getVocabularies}
\alias{getVocabularies}
\title{getVocabularies}
\usage{
getVocabularies(cdm)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}
}
\value{
Names of available vocabularies
}
\description{
getVocabularies
}
\examples{
\dontrun{
cdm <- mockVocabRef()
getVocabularies(cdm = cdm)
CDMConnector::cdmDisconnect(cdm)
}
}
