% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabUtilities.R
\name{getVocabVersion}
\alias{getVocabVersion}
\title{getVocabVersion}
\usage{
getVocabVersion(cdm)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}
}
\value{
the vocabulary version being used
}
\description{
getVocabVersion
}
\examples{
cdm <- mockVocabRef()
getVocabVersion(cdm = cdm)
DBI::dbDisconnect(attr(cdm, "dbcon"), shutdown = TRUE)
}
