% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{keepOriginalDoc}
\alias{keepOriginalDoc}
\title{Helper for consistent documentation of \code{keepOriginal}.}
\arguments{
\item{keepOriginal}{Whether to keep the original codelist (TRUE) or just return
the stratified ones (FALSE).}
}
\description{
Helper for consistent documentation of \code{keepOriginal}.
}
\keyword{internal}
