% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockVocabRef.R
\name{mockVocabRef}
\alias{mockVocabRef}
\title{Generate example vocabulary database}
\usage{
mockVocabRef(backend = "data_frame")
}
\arguments{
\item{backend}{'database' (duckdb) or 'data_frame'.}
}
\value{
cdm reference with mock vocabulary.
}
\description{
Generate example vocabulary database
}
\examples{
\donttest{
library(CodelistGenerator)
cdm <- mockVocabRef()
cdm
}
}
