% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesInUse.R
\name{restrictToCodesInUse}
\alias{restrictToCodesInUse}
\title{Filter a codelist to keep only the codes used in the database}
\usage{
restrictToCodesInUse(
  x,
  cdm,
  table = c("condition_occurrence", "device_exposure", "drug_exposure", "measurement",
    "observation", "procedure_occurrence", "visit_occurrence")
)
}
\arguments{
\item{x}{A codelist}

\item{cdm}{cdm_reference via CDMConnector}

\item{table}{cdm table}
}
\description{
Filter a codelist to keep only the codes used in the database
}
