% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_mappings.R
\name{show_mappings}
\alias{show_mappings}
\title{Show mappings from source vocabularies to standard}
\usage{
show_mappings(
  candidate_codelist,
  source_vocabularies = c("ATC", "ICD10CM", "ICD10PCS", "ICD9CM", "ICD9Proc", "LOINC",
    "OPCS4", "Read", "RxNorm", "RxNorm Extension", "SNOMED"),
  db,
  vocabulary_database_schema
)
}
\arguments{
\item{candidate_codelist}{Dataframe}

\item{source_vocabularies}{Character vector}

\item{db}{Database connection via DBI::dbConnect()}

\item{vocabulary_database_schema}{Name of database schema with vocab tables}
}
\value{
tibble
}
\description{
Show mappings from source vocabularies to standard
}
\examples{
\dontrun{
library(DBI)
library(CodelistGenerator)
db <- DBI::dbConnect(" Your database connection here " )
vocabulary_database_schema <- " Your vocabulary schema here "
asthma_codes <- get_candidate_codes(
  keywords = "asthma",
  db = db,
  vocabulary_database_schema = " Your vocabulary schema here ")
show_mappings(
  candidate_codelist = asthma_codes,
  db = db,
  vocabulary_database_schema = " Your vocabulary schema here "
)
}
}
