% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_mappings.R
\name{show_mappings}
\alias{show_mappings}
\title{Show mappings from source vocabularies to standard}
\usage{
show_mappings(
  candidate_codelist,
  source_vocabularies = c("ATC", "ICD10CM", "ICD10PCS", "ICD9CM", "ICD9Proc", "LOINC",
    "OPCS4", "Read", "RxNorm", "RxNorm Extension", "SNOMED"),
  db,
  vocabulary_database_schema
)
}
\arguments{
\item{candidate_codelist}{Dataframe}

\item{source_vocabularies}{Character vector}

\item{db}{Database connection via DBI::dbConnect()}

\item{vocabulary_database_schema}{Name of database schema with vocab tables}
}
\value{
Dataframe
}
\description{
Show mappings from source vocabularies to standard
}
\examples{
### note, Eunomia, which is used for the example below,
### does not include a full set of vocabularies.
### The full set can be downloaded from https://athena.ohdsi.org
library(Eunomia)
library(RSQLite)
library(dplyr)
library(DBI)
untar(xzfile(system.file("sqlite", "cdm.tar.xz",
  package = "Eunomia"
), open = "rb"),
exdir = tempdir()
)
db <- DBI::dbConnect(RSQLite::SQLite(), paste0(tempdir(), "\\\\cdm.sqlite"))
asthma_codes <- get_candidate_codes(
  keywords = "asthma",
  db = db,
  vocabulary_database_schema = "main"
)
show_mappings(
  candidate_codelist = asthma_codes,
  db = db,
  vocabulary_database_schema = "main"
)
}
