% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMappings.R
\name{showMappings}
\alias{showMappings}
\title{Show mappings from non-standard vocabularies to standard}
\usage{
showMappings(
  candidateCodelist,
  nonStandardVocabularies = c("ATC", "ICD10CM", "ICD10PCS", "ICD9CM", "ICD9Proc",
    "LOINC", "OPCS4", "Read", "RxNorm", "RxNorm Extension", "SNOMED"),
  db,
  vocabularyDatabaseSchema
)
}
\arguments{
\item{candidateCodelist}{Dataframe}

\item{nonStandardVocabularies}{Character vector}

\item{db}{Database connection via DBI::dbConnect()}

\item{vocabularyDatabaseSchema}{Name of database schema with vocab tables}
}
\value{
tibble
}
\description{
Show mappings from non-standard vocabularies to standard
}
\examples{
\dontrun{
library(DBI)
library(CodelistGenerator)
db <- DBI::dbConnect(" Your database connection here ")
vocabularyDatabaseSchema <- " Your vocabulary schema here "
asthma_codes <- get_candidate_codes(
  keywords = "asthma",
  db = db,
  vocabularyDatabaseSchema = " Your vocabulary schema here "
)
showMappings(
  candidateCodelist = asthma_codes,
  db = db,
  vocabularyDatabaseSchema = " Your vocabulary schema here "
)
}
}
