% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesInUse.R
\name{sourceCodesInUse}
\alias{sourceCodesInUse}
\title{Get source codes used in the database}
\usage{
sourceCodesInUse(
  cdm,
  table = c("condition_occurrence", "device_exposure", "drug_exposure", "measurement",
    "observation", "procedure_occurrence", "visit_occurrence")
)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{table}{cdm table}
}
\value{
A list of source codes used in the database.
}
\description{
Get source codes used in the database
}
\examples{
\dontrun{
cdm <- mockVocabRef("database")
x <- sourceCodesInUse(cdm = cdm)
x
DBI::dbDisconnect(attr(cdm, "dbcon"), shutdown = TRUE)
}
}
