% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyByConcept.R
\name{stratifyByConcept}
\alias{stratifyByConcept}
\title{Stratify a codelist by the concepts included within it}
\usage{
stratifyByConcept(x, cdm, keepOriginal = FALSE)
}
\arguments{
\item{x}{A codelist}

\item{cdm}{A cdm reference}

\item{keepOriginal}{Whether to keep the original codelist and append the
stratify (if TRUE) or just return the stratified codelist (if FALSE).}
}
\value{
A codelist
}
\description{
Stratify a codelist by the concepts included within it
}
