% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyByDoseForm.R
\name{stratifyByDoseForm}
\alias{stratifyByDoseForm}
\title{Stratify a codelist by dose form.}
\usage{
stratifyByDoseForm(
  x,
  cdm,
  nameStyle = "{codelist_name}_{dose_form}",
  keepOriginal = FALSE
)
}
\arguments{
\item{x}{A codelist.}

\item{cdm}{A cdm reference to an OMOP CDM dataset. If data is held within a
database, the vocabulary tables should be in the same schema as the clinical
tables (person, observation period, and so on).}

\item{nameStyle}{Naming of the new codelists, use \code{{codelist_name}} to
include the codelist name and \code{{dose_form}} to include the dose form name.}

\item{keepOriginal}{Whether to keep the original codelist (TRUE) or just return
the stratified ones (FALSE).}
}
\value{
The codelist with the required stratifications, as different elements
of the list.
}
\description{
Stratify a codelist by dose form.
}
\examples{
\donttest{
library(CodelistGenerator)

cdm <- mockVocabRef()

codes <- newCodelist(list("codes" = c(10L, 20L, 21L)))
new_codes <- stratifyByDoseForm(x = codes,
                                cdm = cdm,
                                keepOriginal = TRUE)
new_codes
}
}
