% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyByRoute.R
\name{stratifyByRouteCategory}
\alias{stratifyByRouteCategory}
\title{Stratify a codelist by route category}
\usage{
stratifyByRouteCategory(x, cdm)
}
\arguments{
\item{x}{A codelist}

\item{cdm}{A cdm reference}
}
\value{
A codelist
}
\description{
Stratify a codelist by route category
}
