% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyByRoute.R
\name{stratifyByRouteCategory}
\alias{stratifyByRouteCategory}
\title{Stratify a codelist by route category}
\usage{
stratifyByRouteCategory(x, cdm, remove = TRUE)
}
\arguments{
\item{x}{A codelist}

\item{cdm}{A cdm reference}

\item{remove}{TRUE or FALSE. If TRUE the inputted codelist will be kept and
the stratified appended. If FALSE only the stratified codelist will be
returned.}
}
\value{
A codelist
}
\description{
Stratify a codelist by route category
}
