% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetOnDomain.R
\name{subsetOnDomain}
\alias{subsetOnDomain}
\title{Subset a codelist to only those codes from a particular domain}
\usage{
subsetOnDomain(x, cdm, domain)
}
\arguments{
\item{x}{Codelist}

\item{cdm}{A cdm reference}

\item{domain}{Domains. Use getDomains() to find the available
domains in a cdm}
}
\value{
The codelist with only those concepts associated with the
domain
}
\description{
Subset a codelist to only those codes from a particular domain
}
