% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetOnDoseUnit.R
\name{subsetOnDoseUnit}
\alias{subsetOnDoseUnit}
\title{Subset a codelist to only those with a particular dose unit}
\usage{
subsetOnDoseUnit(x, cdm, doseUnit)
}
\arguments{
\item{x}{Codelist}

\item{cdm}{A cdm reference}

\item{doseUnit}{Dose unit. Use getDoseUnit() to find the available
dose units in a cdm}
}
\value{
The codelist with only those concepts associated with the
dose unit
}
\description{
Subset a codelist to only those with a particular dose unit
}
