% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetOnRouteCategory.R
\name{subsetOnRouteCategory}
\alias{subsetOnRouteCategory}
\title{Subset a codelist to only those with a particular route category}
\usage{
subsetOnRouteCategory(x, cdm, routeCategory)
}
\arguments{
\item{x}{Codelist}

\item{cdm}{A cdm reference}

\item{routeCategory}{Route category. Use getRoutes() to find the available
route categories for a cdm}
}
\value{
The codelist with only those concepts associated with the
specified route categories
}
\description{
Subset a codelist to only those with a particular route category
}
