% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseUnmappedCodes.R
\name{summariseUnmappedCodes}
\alias{summariseUnmappedCodes}
\title{Findunmapped concepts related to codelist}
\usage{
summariseUnmappedCodes(
  x,
  cdm,
  table = c("condition_occurrence", "device_exposure", "drug_exposure", "measurement",
    "observation", "procedure_occurrence")
)
}
\arguments{
\item{x}{A codelist}

\item{cdm}{A cdm reference}

\item{table}{Names of clinical tables in which to search for unmapped codes.
Can be one or more of "condition_occurrence", "device_exposure",
"drug_exposure", "measurement", "observation", and "procedure_occurrence".}
}
\value{
A summarised result
}
\description{
Findunmapped concepts related to codelist
}
