% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{tableStyleDoc}
\alias{tableStyleDoc}
\title{Helper for consistent documentation of \code{style}.}
\arguments{
\item{style}{A character string or custom R code to define the visual formatting
of the table. This argument can be provided in two ways: (1) Pre-defined Styles
(Character String): Use a name for a built-in style (e.g., "darwin"). See
visOmopResults::tableStyle() for available options. (2) Custome Code (Advanced):
Supply a block of custom R code. This code must be specific to the table type.
See visOmopResults::tableStyleCode() for structural examples.}
}
\description{
Helper for consistent documentation of \code{style}.
}
\keyword{internal}
