% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_helper.R
\name{xDocCohort}
\alias{xDocCohort}
\title{Helper for consistent documentation of \code{x} where input can be codelist or cohort.}
\arguments{
\item{x}{A codelist or cohort table name.}
}
\description{
Helper for consistent documentation of \code{x} where input can be codelist or cohort.
}
\keyword{internal}
